Refactor AI module into separate files
