+++ README.md
@@ -2,5 +2,5 @@
 This repository contains the game code and the AI module.
 
 ## Structure
-- `src/UltraWorldAI/IA.cs` contains the main AI implementation.
+- `src/UltraWorldAI/` now contains separate files for each AI class.
 - `AIConfig.json` can override some runtime parameters.
