+++ src/UltraWorldAI/StressSystem.cs
@@ -1,0 +1,40 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.IO;
+using System.Text.Json;
+
+namespace UltraWorldAI
+{
+    public class StressSystem
+    {
+        public float CurrentStressLevel { get; private set; }
+        private readonly Person _person;
+
+        public StressSystem(Person person)
+        {
+            _person = person;
+            CurrentStressLevel = AIConfig.MinStress;
+        }
+
+        public void AddStress(float amount)
+        {
+            CurrentStressLevel = Math.Min(AIConfig.MaxStress, CurrentStressLevel + amount);
+        }
+
+        public void ReduceStress(float amount)
+        {
+            CurrentStressLevel = Math.Max(AIConfig.MinStress, CurrentStressLevel - amount);
+        }
+
+        public void UpdateStressDecay()
+        {
+            ReduceStress(AIConfig.StressDecayRate);
+        }
+
+        public bool IsStressed()
+        {
+            return CurrentStressLevel > 0.5f;
+        }
+    }
+}
